projects/Chat-GPT-Discord-Bot/requirements.txt
discord.py~=2.3.2
openai~=1.30.1
python-dotenv~=1.0.1